+++ state/interface.go
@@ -18,23 +18,11 @@ type EntityFinder interface {
 	FindEntity(tag names.Tag) (Entity, error)
 }
 
-var _ EntityFinder = (*State)(nil)
-
 // Entity represents any entity that can be returned
 // by State.FindEntity. All entities have a tag.
 type Entity interface {
 	Tag() names.Tag
 }
-
-var (
-	_ Entity = (*Machine)(nil)
-	_ Entity = (*Unit)(nil)
-	_ Entity = (*UnitAgent)(nil)
-	_ Entity = (*Service)(nil)
-	_ Entity = (*Environment)(nil)
-	_ Entity = (*User)(nil)
-	_ Entity = (*Action)(nil)
-)
 
 // EntityWithService is implemented by Units it is intended
 // for anything that can return its Service.
@@ -42,19 +30,10 @@ type EntityWithService interface {
 	Service() (*Service, error)
 }
 
-var _ EntityWithService = (*Unit)(nil)
-
 // Lifer represents an entity with a life.
 type Lifer interface {
 	Life() Life
 }
-
-var (
-	_ Lifer = (*Machine)(nil)
-	_ Lifer = (*Unit)(nil)
-	_ Lifer = (*Service)(nil)
-	_ Lifer = (*Relation)(nil)
-)
 
 // AgentTooler is implemented by entities
 // that have associated agent tools.
@@ -68,20 +47,10 @@ type EnsureDeader interface {
 	EnsureDead() error
 }
 
-var (
-	_ EnsureDeader = (*Machine)(nil)
-	_ EnsureDeader = (*Unit)(nil)
-)
-
 // Remover represents entities with a Remove method.
 type Remover interface {
 	Remove() error
 }
-
-var (
-	_ Remover = (*Machine)(nil)
-	_ Remover = (*Unit)(nil)
-)
 
 // Authenticator represents entites capable of handling password
 // authentication.
@@ -91,24 +60,11 @@ type Authenticator interface {
 	PasswordValid(pass string) bool
 }
 
-var (
-	_ Authenticator = (*Machine)(nil)
-	_ Authenticator = (*Unit)(nil)
-	_ Authenticator = (*User)(nil)
-)
-
 // NotifyWatcherFactory represents an entity that
 // can be watched.
 type NotifyWatcherFactory interface {
 	Watch() NotifyWatcher
 }
-
-var (
-	_ NotifyWatcherFactory = (*Machine)(nil)
-	_ NotifyWatcherFactory = (*Unit)(nil)
-	_ NotifyWatcherFactory = (*Service)(nil)
-	_ NotifyWatcherFactory = (*Environment)(nil)
-)
 
 // AgentEntity represents an entity that can
 // have an agent responsible for it.
@@ -123,19 +79,12 @@ type AgentEntity interface {
 	NotifyWatcherFactory
 }
 
-var (
-	_ AgentEntity = (*Machine)(nil)
-	_ AgentEntity = (*Unit)(nil)
-)
-
 // EnvironAccessor defines the methods needed to watch for environment
 // config changes, and read the environment config.
 type EnvironAccessor interface {
 	WatchForEnvironConfigChanges() NotifyWatcher
 	EnvironConfig() (*config.Config, error)
 }
-
-var _ EnvironAccessor = (*State)(nil)
 
 // UnitsWatcher defines the methods needed to retrieve an entity (a
 // machine or a service) and watch its units.
@@ -144,23 +93,16 @@ type UnitsWatcher interface {
 	WatchUnits() StringsWatcher
 }
 
-var _ UnitsWatcher = (*Machine)(nil)
-var _ UnitsWatcher = (*Service)(nil)
-
 // EnvironMachinesWatcher defines a single method -
 // WatchEnvironMachines.
 type EnvironMachinesWatcher interface {
 	WatchEnvironMachines() StringsWatcher
 }
 
-var _ EnvironMachinesWatcher = (*State)(nil)
-
 // InstanceIdGetter defines a single method - InstanceId.
 type InstanceIdGetter interface {
 	InstanceId() (instance.Id, error)
 }
-
-var _ InstanceIdGetter = (*Machine)(nil)
 
 // ActionsWatcher defines the methods an entity exposes to watch Actions
 // queued up for itself
@@ -168,12 +110,6 @@ type ActionsWatcher interface {
 	Entity
 	WatchActionNotifications() StringsWatcher
 }
-
-var (
-	_ ActionsWatcher = (*Unit)(nil)
-	// TODO(jcw4): when we implement service level Actions
-	// _ ActionsWatcher = (*Service)(nil)
-)
 
 // ActionReceiver describes Entities that can have Actions queued for
 // them, and that can get ActionRelated information about those actions.
@@ -212,22 +148,8 @@ type ActionReceiver interface {
 	RunningActions() ([]*Action, error)
 }
 
-var (
-	_ ActionReceiver = (*Unit)(nil)
-	// TODO(jcw4) - use when Actions can be queued for Services.
-	//_ ActionReceiver = (*Service)(nil)
-)
-
 // GlobalEntity specifies entity.
 type GlobalEntity interface {
 	globalKey() string
 	Tag() names.Tag
 }
-
-var (
-	_ GlobalEntity = (*Machine)(nil)
-	_ GlobalEntity = (*Unit)(nil)
-	_ GlobalEntity = (*Service)(nil)
-	_ GlobalEntity = (*Charm)(nil)
-	_ GlobalEntity = (*Environment)(nil)
-)
