+++ state/open.go
@@ -172,6 +172,7 @@ var indexes = []struct {
 	{networkInterfacesC, []string{"env-uuid", "machineid"}, false, false},
 	{blockDevicesC, []string{"env-uuid", "machineid"}, false, false},
 	{subnetsC, []string{"providerid"}, true, true},
+	{ipaddressesC, []string{"ipaddress-uuid"}, false, false},
 	{ipaddressesC, []string{"env-uuid", "state"}, false, false},
 	{ipaddressesC, []string{"env-uuid", "subnetid"}, false, false},
 	{storageInstancesC, []string{"env-uuid", "owner"}, false, false},
